Fix subtle bug in calculation of possible telemarketers
The set of called numbers has to be complete and stable before the
set of possible telemarketers can be compared to it. The previous
implementation was wrong because it checked on each record whether
the calling number had been called yet, which ignores the possibility
that the number is called later on in the dataset. The new solution
uses a set difference to remove the numbers that have received calls
from the list of possible telemarketers after the initial filtering
of numbers that are in the texting logs has been done.